agent_service/requirements.txt
google-adk>=1.25.0
fastapi>=0.115.12
uvicorn[standard]>=0.34.0
firebase-admin>=6.8.0
python-dotenv>=1.0.1
